% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{axes}
\alias{axes}
\alias{axes}
\alias{axes.default}
\title{Collection of axes.}
\usage{
axes(...)

\method{axes}{default}(...)
}
\arguments{
\item{...}{d_axis inputs}
}
\description{
Collection of axes.
}
