% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{d_axis}
\alias{d_axis}
\alias{d_axis}
\alias{d_axis.default}
\title{Abstract axis}
\usage{
d_axis(n, min, max, name)

\method{d_axis}{default}(n, min = 0, max = n, name = NA_character_)
}
\arguments{
\item{n}{number of discrete steps in the axis}

\item{min}{the minimum location in the axis (defaults to zero)}

\item{max}{the maximum location in the axis (defaults to `n`)}

\item{name}{a name to apply to the axis}
}
\description{
Create an axis from inputs `n`umber, `min`imum and `max`imum, and optionally
a `name`. The name is intended to be present and unique for downstream use, but in
isolation are optional.
}
\details{
Continuous axes are supported, but only for some kind of completeness. The
intention is for straightforward 1-D discretization and the ability to be continuous is
probably better supported elsewhere (i.e. the scales package).
}
