% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_discrete.R
\name{dc_discrete}
\alias{dc_discrete}
\title{Discrete}
\usage{
dc_discrete(x, ..., mins = NULL, maxs = NULL, dims = NULL, name = NULL)
}
\arguments{
\item{x}{object to derive properties from (all other named arguments are ignored)}

\item{...}{ignored currently}

\item{mins}{minimum coordinates}

\item{maxs}{maximum coordinates}

\item{dims}{axis length}

\item{name}{optional name for each axis}
}
\description{
table of axes
}
