% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiling-axes.R
\name{disc_rete}
\alias{disc_rete}
\title{table of axes}
\usage{
disc_rete(mins, maxs, dims, name = NULL)
}
\arguments{
\item{mins}{minimum coordinates}

\item{maxs}{maximum coordinates}

\item{dims}{axis length}

\item{name}{optional name for each axis}
}
\description{
table of axes
}
