% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiling-axes.R
\name{gdal_source}
\alias{gdal_source}
\title{the index extent of a local geom () to  a real extent of `obj` converted to GDAL's standard
offsetX/Y, srcwinX/Y form}
\usage{
gdal_source(obj, geom)
}
\description{
the index extent of a local geom () to  a real extent of `obj` converted to GDAL's standard
offsetX/Y, srcwinX/Y form
}
