% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiling-axes.R
\name{pair_four}
\alias{pair_four}
\title{the modulo "n-column" index  required
for the local 4-cornes of a cell from the
ordered vertex pool}
\usage{
pair_four(xp, nc)
}
\description{
the modulo "n-column" index  required
for the local 4-cornes of a cell from the
ordered vertex pool
}
