% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiling-axes.R
\name{populate_raster}
\alias{populate_raster}
\title{worker to build an empty list of NULLs, except optionally
for one of the cells to hold a table of data}
\usage{
populate_raster(n, tile = NULL, data = NULL)
}
\description{
worker to build an empty list of NULLs, except optionally
for one of the cells to hold a table of data
}
