% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiling-axes.R
\name{tile_gon}
\alias{tile_gon}
\title{fast creation of simple features tiles from abstract
axis specification
methods? raster_info, geotransform/dim, raster}
\usage{
tile_gon(x, crs = NA_character_, ...)
}
\description{
fast creation of simple features tiles from abstract
axis specification
methods? raster_info, geotransform/dim, raster
}
