% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiling-axes.R
\name{tile_heuristic}
\alias{tile_heuristic}
\title{tile guess based on overall dimension}
\usage{
tile_heuristic(dim, base_guess = c(1024L, 1024L), fix = FALSE)
}
\description{
tile guess based on overall dimension
}
