% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiling-axes.R
\name{transform6}
\alias{transform6}
\title{6 figure transform to discrete axes}
\usage{
transform6(x, dim)
}
\arguments{
\item{x}{6 figure geotransform as returned by raster_info}

\item{dim}{2 figure dimXY as returned by raster_info}
}
\description{
6 figure transform to discrete axes
}
