% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiling-axes.R
\name{dummy_discrete}
\alias{dummy_discrete}
\title{a default discretization for testing, called by `disc_rete`}
\usage{
dummy_discrete()
}
\description{
a default discretization for testing, called by `disc_rete`
}
